changing the functions in the startupcontoller

+++ database/migrations/2024_09_04_192649_create_startup_sectors_table.php
@@ -9,8 +9,9 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
-    {
+  public function up(): void
+{
+    if (!Schema::hasTable('startup_sectors')) {
         Schema::create('startup_sectors', function (Blueprint $table) {
             $table->id();
             $table->foreignId('startup_id')->constrained('startups')->onDelete('cascade');
@@ -18,10 +19,9 @@ return new class extends Migration
             $table->foreign('sector_id')->references('id')->on('sectors');
         });
     }
+}
 
-    /**
-     * Reverse the migrations.
-     */
+  
     public function down(): void
     {
         Schema::dropIfExists('startup_sectors');
